maf correction update

+++ src/gaf_creator.hpp
@@ -28,12 +28,12 @@ using namespace std;
 
 namespace stoat_vcf {
 
-std::pair<double, double> calcul_proportion_signi(int number_ind_group0, int number_ind_group1, double p_value);
+std::pair<double, double> calcul_proportion_signi(size_t number_ind_group0, size_t number_ind_group1, double p_value);
 std::string addSuffixToFilename(const std::string& filename, const std::string& suffix);
 std::string add_suffix_to_filename(const std::string& filename, const std::string& suffix);
 void write_gaf_lines(const std::string& sequence_name, const std::string& path, int length, double prop, ofstream& outfile);
 
-std::vector<int> decompose_snarl(const std::string& snarl);
+std::vector<size_t> decompose_snarl(const std::string& snarl);
 int calcul_path_length(bdsg::PackedGraph& pg, const std::string& snarl);
 void write_gaf_lines(const std::string& sequence_name, const std::string& path, int length, double prop, ofstream& outfile);
 void gaf_creation(const std::string& input_file, std::unordered_map<std::string, std::vector<stoat_vcf::Snarl_data_t>>& snarl_chr,
